Add method to perform POST for calls that require it
Some of the Untappd API calls - Checkin for example - require that POST
be used instead of GET.

This adds a new method - _sendPost - that mirrors most of the
functionality within _request,   takes in the same parameters and
outputs the same response.  The only difference is that it does a POST
instead of a GET.

I also updated the methods that - according to the API Documentation -
need to be sent via POST to use the new method.

I've tested this working successfully for checkins, where the _request
method resulted in 500 errors.